feat: refactory & add tests
