+++ .pre-commit-config.yaml
@@ -45,10 +45,10 @@ repos:
         language: rust
         files: \.rs$
         pass_filenames: false
-      # - id: cargo-test
-      #   name: cargo test
-      #   description: unit test for the project
-      #   entry: bash -c 'cargo test --all-features --all'
-      #   language: rust
-      #   files: \.rs$
-      #   pass_filenames: false
+      - id: cargo-test
+        name: cargo test
+        description: unit test for the project
+        entry: bash -c 'cargo test --all-features --all'
+        language: rust
+        files: \.rs$
+        pass_filenames: false
