+++ src/stream/mod.rs
@@ -1,5 +1,7 @@
+mod gst_pipeline;
 mod pipeline;
 mod utils;
 
+pub use gst_pipeline::*;
 pub use pipeline::*;
 pub use utils::*;
